Updated references to renamed generateOutageChart.js (now chartFunctions.js)
