+++ backend/chartFunctions.js
